Implement Warning the user after successful added if the series was 100 before to reach the maximum 10 digits SeriesNumber in SOA

+++ Repository/StatementOfAccountRepo.cs
@@ -42,7 +42,7 @@ namespace Accounting_System.Repository
         //    }
         //}
 
-        public async Task<int> GetLastSeriesNumber()
+        public async Task<long> GetLastSeriesNumber()
         {
             var lastInvoice = await _dbContext
                 .StatementOfAccounts
